requirements.txt
haversine==2.8.1
mysql-connector-python==8.0.33
tabulate==0.9.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
python-dotenv==1.0.0